+++ src/Console/Components/Templates/Controller.php
@@ -6,11 +6,15 @@ use Solital\Core\Course\Container\Container;
 use Solital\Core\Resource\Message;
 
 /**
+ * 
+ * DO NOT MODIFY THIS CONTROLLER --------------------------------------------------------------------
+ * 
+ * This is a generic class that can be changed at any update. 
+ *
  * This class is a generic Controller, all Controllers created in Solital must extend this class. 
  * 
  * It offers ways to access some classes (such as Message and Container) in a simpler and easier way. 
  */
-
 abstract class Controller
 {
     /**